+++ trueseeing/signature/privacy.py
@@ -30,7 +30,7 @@ from typing import TYPE_CHECKING
 
 import re
 
-from trueseeing.core.flow.code import InvocationPattern
+from trueseeing.core.code.model import InvocationPattern
 from trueseeing.core.flow.data import DataFlows
 from trueseeing.signature.base import Detector
 from trueseeing.core.issue import Issue
