+++ trueseeing/signature/security.py
@@ -31,7 +31,7 @@ import lxml.etree as ET
 import re
 import os
 
-from trueseeing.core.flow.code import InvocationPattern
+from trueseeing.core.code.model import InvocationPattern
 from trueseeing.core.flow.data import DataFlows
 from trueseeing.signature.base import Detector
 from trueseeing.core.issue import Issue
